# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mi Dulce Compania - Backoffice API",
    "description": "Contrato del backend para clientes web. UUID como string; fechas ISO-8601 UTC.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Servidor actual"
    }
  ],
...
  "paths": {
...
    "/api/v1/materials/{id}": {
      "get": {
        "tags": [
          "Materials"
        ],
        "summary": "Consultar por UUID",
        "operationId": "getMaterialById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Peticion invalida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de integridad referencial",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "200": {
            "description": "Recurso encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Materials"
        ],
        "summary": "Actualizar recurso",
        "operationId": "updateMaterial",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Peticion invalida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de integridad referencial",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "200": {
            "description": "Recurso actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Materials"
        ],
        "summary": "Eliminar recurso",
        "operationId": "deleteMaterial",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Peticion invalida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de integridad referencial",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "204": {
            "description": "Recurso eliminado"
          }
        }
      }
    },
...
    "/api/v1/materials": {
      "get": {
        "tags": [
          "Materials"
        ],
        "summary": "Listar con paginacion",
        "operationId": "listMaterials",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [
                "name,ASC"
              ]
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Peticion invalida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de integridad referencial",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "200": {
            "description": "Pagina de resultados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDTOMaterialResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Materials"
        ],
        "summary": "Crear recurso",
        "operationId": "createMaterial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Peticion invalida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de integridad referencial",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "201": {
            "description": "Recurso creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiProblem": {
        "required": [
          "detail",
          "instance",
          "status",
          "title",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "urn:problem:http:400"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/cities"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "description": "Mapa campo -> mensaje; presente solo para errores de validacion."
            },
            "description": "Mapa campo -> mensaje; presente solo para errores de validacion."
          }
        },
        "description": "Error ProblemDetail. errors contiene mensajes por campo en errores de validacion."
      },
...
      "MaterialRequestDTO": {
        "required": [
          "currentStock",
          "name",
          "unitOfMeasure"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string"
          },
          "unitOfMeasure": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "currentStock": {
            "minimum": 0,
            "type": "number"
          }
        },
        "description": "Material request"
      },
      "MaterialResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "unitOfMeasure": {
            "type": "string"
          },
          "currentStock": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string"
          }
        },
        "description": "Material response"
      },
...
      "PageResponseDTOMaterialResponseDTO": {
        "required": [
          "content",
          "first",
          "last",
          "number",
          "size",
          "totalElements",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MaterialResponseDTO"
            }
          },
          "number": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        },
        "description": "Pagina estable del contrato REST; la primera pagina tiene numero 0."
      },
...
    }
  }
}